+++ src/pygpt_net/core/events/kernel.py
@@ -6,7 +6,7 @@
 # GitHub:  https://github.com/szczyglis-dev/py-gpt   #
 # MIT License                                        #
 # Created By  : Marcin Szczygliński                  #
-# Updated Date: 2024.12.14 08:00:00                  #
+# Updated Date: 2025.07.25 22:00:00                  #
 # ================================================== #
 
 from typing import Optional
@@ -35,6 +35,7 @@ class KernelEvent(BaseEvent):
     APPEND_END = "kernel.append.end"
     CALL = "kernel.call"
     REQUEST = "kernel.request"
+    REQUEST_CALL = "kernel.request_call"
     REQUEST_NEXT = "kernel.request.next"
     RESPONSE_ERROR = "kernel.response.error"
     RESPONSE_FAILED = "kernel.response.failed"