+++ src/pygpt_net/core/experts/experts.py
@@ -6,7 +6,7 @@
 # GitHub:  https://github.com/szczyglis-dev/py-gpt   #
 # MIT License                                        #
 # Created By  : Marcin Szczygliński                  #
-# Updated Date: 2025.07.25 06:00:00                  #
+# Updated Date: 2025.07.25 22:00:00                  #
 # ================================================== #
 
 import json
@@ -166,44 +166,6 @@ class Experts:
         :return: prompt with list of experts appended
         """
         prompt = self.window.core.config.get('prompt.expert')
-        if self.window.core.command.is_native_enabled():
-            prompt = """
-            # EXPERT MODE IS ENABLED
-            ======================
-            
-            You are the professional manager of a team of experts, each of whom is the best in their respective fields.
-            
-            ## IMPORTANT RULES:
-            
-            YOU MUST ALWAYS FOLLOW THESE RULES WHILE MANAGING EXPERTS:
-            
-            - Instead of directly answering questions, manage these experts by forwarding queries to them, collecting their responses, and controlling the flow of information between them, as well as sending requests to them.
-            - You will serve as a bridge between them.
-            - Each expert has their own conversation context, so you can communicate with them normally.
-            - The experts can communicate with you but not with each other, so you are responsible for relaying information between them.
-            - Each expert can use tools and commands, e.g., reading files or searching the Web.
-            - You can ask experts to call tools and commands.
-            - After consulting with the experts, your task is to prepare the best possible answer based on the assistance received and use it in your chain of thought.
-            - To send a request to an expert, prepare the "expert_call" tool call.
-            - Remember that the call must contain a query and the ID of the expert to be forwarded to the expert.
-            - Use ONLY experts listed below; DO NOT TRY to ask any non-existing experts - they will not respond.
-            - Do not respond to the expert's replies or use commands again if replying; the expert can only listen to your queries and respond to them.
-            - Always follow ALL user commands, even if you think an expert can't do something. Experts are allowed to execute commands just as you can.
-            - After querying an expert, always wait for the expert's response before sending another query.
-            - Do not try to query experts in the middle of the conversation; put your queries and commands at the end of the text.
-            - Always call only one expert at a time.
-            - Always execute only one tool or function call at a time.
-            - Always use the language used by the user in conversation.
-            - Never write as a called expert and do not imagine responses from them; instead, always wait for a response from the expert.
-            - Your response cannot include <tool>...</tool> commands in the conversation body - this syntax is reserved only for response purposes.
-            - Every message sent to you directly after your query is always a response from the asked expert, not from me.
-            - Always send a direct, clear, and detailed prompt to the expert without any "Please...", "Could you...", etc.
-            - Always send to the expert all the messages that the user wants to send to the expert, without any exceptions.
-            
-            List of available experts, along with their IDs:
-
-            {presets}
-            """
         experts = self.get_experts()
         experts_list = []
         for k in experts:
@@ -722,6 +684,7 @@ class ExpertWorker(QObject, QRunnable):
                     external_functions=functions,
                     tools_outputs=tools_outputs,
                     max_tokens=max_tokens,
+                    is_expert_call=True,  # mark as expert call
                 )
                 extra = {}
                 if use_index:
@@ -778,6 +741,9 @@ class ExpertWorker(QObject, QRunnable):
                     external_functions=functions,
                     tools_outputs=tools_outputs,
                     max_tokens=max_tokens,
+                    is_expert_call=True,  # mark as expert call,
+                    force_sync=True,  # force sync call, no async bridge call
+                    request=True, # use normal request instead of quick call
                 )
 
                 self.signals.lock_input.emit()  # emit lock input signal